Refactors rjwt-py into separate modules
